Move VendorsTable in components folder and delete tables folder
