+++ src/pages/Vendors.tsx
@@ -5,7 +5,7 @@ import { useTranslation } from 'react-i18next'
 import { TextField, Checkbox, FormControlLabel } from '@mui/material'
 import { Controller, UseFormReturn, SubmitHandler } from 'react-hook-form'
 import { VendorFormData, vendorSchema } from '@/schemas/vendorSchema'
-import VendorsTable from '@/tables/VendorsTable'
+import VendorsTable from '@/components/features/VendorsTable'
 
 export default function Vendors() {
   const { t: translate } = useTranslation()
